feat: планеты/осада, альянсы, миграции 031-035, снабженец вне состава флота, UI и баланс (GAME_VERSION 01.189)

+++ server/app/db/models/reserved_display_name.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import uuid
 from datetime import datetime
 
-from sqlalchemy import BigInteger, DateTime, ForeignKey, String
+from sqlalchemy import DateTime, ForeignKey, Integer, String
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import Mapped, mapped_column
 
@@ -15,7 +15,8 @@ class ReservedDisplayName(Base):
 
     __tablename__ = "reserved_display_names"
 
-    id: Mapped[int] = mapped_column(BigInteger(), primary_key=True, autoincrement=True)
+    # Integer PK: SQLite корректно делает AUTOINCREMENT; BigInteger без INTEGER PRIMARY KEY давал NULL id.
+    id: Mapped[int] = mapped_column(Integer(), primary_key=True, autoincrement=True)
     name_norm: Mapped[str] = mapped_column(String(64), nullable=False, unique=True, index=True)
     display_snapshot: Mapped[str] = mapped_column(String(64), nullable=False)
     player_id: Mapped[uuid.UUID | None] = mapped_column(
